Porting openshift/cluster-logging-operator rsyslog configuration into linux-system-roles.
Along with this effort, cluster-logging-operator is being updated to
generate rsyslog configuration files from linux-system-roles/logging.
The generated files are placed in the working directory once and will
be deployed in the containers.  To support the ability to place files
in a directory which is different from the final location, additional
config_dir rsyslog_file_config_dir has been introduced.
1) rsyslog_config_dir for the path in the config files, and
2) rsyslog_file_config_dir for the path to place the config files.

New variables:
- `rsyslog_parent_config_dir`: Directory to store rsyslog.conf.
  Default to '/etc'.
- `rsyslog_files_config_dir`: Directory to locate rsyslog.conf.
  Used in deploying the config files.  Default to '/etc/rsyslog.d'.
- `rsyslog_mode`:  Mode for the config files.  Default to '0400'
- `rsyslog_max_message_size`: Maximum supported message size.
  Default to 8k.
- `use_rsyslog_image`: If set to 'True', the rsyslog is run in the
  OpenShift container.

Rsyslog allows to use environment variables as string constants as follows.
  key=`echo $ENVVAR`
It is used in the script to start the rsyslog container.  They are
activated only when use_rsyslog_image is true.
  {% if use_rsyslog_image|bool %}
    key=`echo $ENVVAR`
  {% endif %}

Another requirement is the include order. It turned out it is crucial
to keep the include order of the configuration files.  To support it,
'format' type is introduced, which sets the prefix 60- to the filename.
To maintain the order more precisely, some files have the secondary
prefix, e.g., 00-elasticsearch.  With the type 'format', it generates
60-00-elasticsearch.conf which is included prior to 60-10-mmk8s.conf.